requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
constantly==15.1.0
cryptography==36.0.2
daphne==3.0.2
Django==3.2.9
dlib==19.18.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
imutils==0.5.4
incremental==21.3.0
msgpack==1.0.3
numpy==1.22.4
opencv-python==4.6.0.66
Pillow==9.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pytz==2021.3
scipy==1.8.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.2.0
txaio==22.2.1
typing_extensions==4.1.1
zope.interface==5.4.0
